migrate tree to separate component
