+++ packages/client/src/components/advanced/EntityTooltip/EntityTooltipStyles.tsx
@@ -43,54 +43,6 @@ export const StyledRelationTypeBlock = styled.div`
   margin-bottom: ${({ theme }) => theme.space[5]};
   /* border: 1px dashed blue; */
 `;
-const getBlockWidth = (depth?: number) => {
-  switch (depth) {
-    case 1:
-      return "10rem";
-    case 2:
-      return "20rem";
-    case 3:
-      return "30rem";
-    default:
-      return "30rem";
-  }
-};
-interface StyledRelationTypeTreeBlock {
-  depth?: number;
-}
-export const StyledRelationTypeTreeBlock = styled.div<StyledRelationTypeTreeBlock>`
-  display: flex;
-  flex-direction: column;
-  margin-left: ${({ theme }) => theme.space[2]};
-  margin-bottom: ${({ theme }) => theme.space[5]};
-  border: 1px dashed white;
-  width: ${({ depth }) => getBlockWidth(depth)};
-`;
-export const StyledTreeBlock = styled.div`
-  border-bottom: 1px solid white;
-  width: 100%;
-  height: 100%;
-`;
-export const StyledFlexColumn = styled.div`
-  display: flex;
-  flex-direction: column;
-  /* TODO: solve thirds */
-`;
-export const StyledFlexRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-`;
-export const StyledGridRowThird = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 2fr;
-  width: 100%;
-`;
-export const StyledGridRowHalf = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  width: 100%;
-`;
 
 export const StyledLetterIconWrap = styled.div`
   display: grid;
